\name{matrixPi1}
\alias{matrixPi1}
\docType{data}
\title{The bernoulli probability matrix for matrixZ1}
\description{The matrix with the bernoulli probability for 
        binary matrixZ1. Therefore, 
        matrixPi1[g,k]=P(matrixZ1[g,k]==1).
        Dim(matrixPi1)=G1*K. This matrix is determined 
        by matrixL1 and eta0, eta1.}
\usage{data(matrixPi1)}
\format{
  The format is:
 num [1:30, 1:10] 0.2 0.2 0.2 0.2 0.2 0.8 0.2 0.2 0.2 0.2 ...
}
\examples{
data(matrixPi1)
}
\keyword{datasets}
