\name{matrixPi2}
\alias{matrixPi2}
\docType{data}
\title{The bernoulli probability matrix for matrixZ2}
\description{The matrix with the bernoulli probability for 
        binary matrixZ2. Therefore, 
        matrixPi2[g,k]=P(matrixZ2[g,k]==1).
        Dim(matrixPi2)=G2*K. This matrix is determined 
        by matrixL2 and eta0, eta1.}
\usage{data(matrixPi2)}
\format{
  The format is:
 num [1:30, 1:10] 0.2 0.2 0.8 0.2 0.8 0.2 0.2 0.2 0.2 0.2 ...
}
\examples{
data(matrixPi2)
}
\keyword{datasets}
