\name{matrixW1}
\alias{matrixW1}
\docType{data}
\title{The factor loading matrix representing the gene-pathway association}
\description{The factor loading matrix representing the degree of 
        influence of the latent factors on individual genes. Dim(matrixW1)=G1*K; 
        Real continuous value with mean 0 and SD 1.}
\usage{data(matrixW1)}
\format{
  The format is:
 num [1:30, 1:10] 0 0 1.76 0 0 ...
}
\examples{
data(matrixW1)
}
\keyword{datasets}
