\name{matrixW2}
\alias{matrixW2}
\docType{data}
\title{The factor loading matrix representing the drug-pathway association}
\description{The factor loading matrix representing the degree of 
        influence of the latent factors on individual drugs. Dim(matrixW2)=G2*K; 
        Real continuous value with mean 0 and SD 1.}
\usage{data(matrixW2)}
\format{
  The format is:
 num [1:30, 1:10] 0 0 0 2.03 -0.744 ...
}
\examples{
data(matrixW2)
}
\keyword{datasets}
