\name{matrixW_chain}
\alias{matrixW_chain}
\docType{data}
\title{The updated matrixW during the Gibbs sampling}
\description{The updated matrixW in each recorded iteration, with 
        format similar to matrixZ_chain.}
\usage{data(matrixW_chain)}
\format{
  The format is:
List of 2
 $ : num [1:5, 1:300] 0 0 0 0.348 0 ...
 $ : num [1:5, 1:300] -0.651 -0.868 -1.013 -0.868 -0.938 ...
}
\examples{
data(matrixW_chain)
}
\keyword{datasets}
