\name{matrixX}
\alias{matrixX}
\docType{data}
\title{The factor activity matrix}
\description{The factor activity matrix with dimension K*J. matrixX[k,j] 
        is the activity value of the k-th latent factor (e.g., pathway) in the j-th
        sample (e.g., cell line). Real continuous value with mean 0 and SD 1.}
\usage{data(matrixX)}
\format{
  The format is:
 num [1:10, 1:15] -0.647 0.405 -0.824 1.463 0.998 ...
}
\examples{
data(matrixX)
}
\keyword{datasets}
