\name{matrixX_chain}
\alias{matrixX_chain}
\docType{data}
\title{The updated matrixX in the Gibbs sampling process}
\description{The updated matrixX in each recorded iteration. An A*B 
        matrix with each row corresponding to the vectorized matrixX in each
        recorded iteration.}
\usage{data(matrixX_chain)}
\format{
  The format is:
 num [1:5, 1:150] 0.1878 0.2226 0.1325 0.0274 -0.1524 ...
}
\examples{
data(matrixX_chain)
}
\keyword{datasets}
