\name{matrixY1}
\alias{matrixY1}
\docType{data}
\title{The gene expression dataset}
\description{The gene expression matrix, dim(Y1)=G1*J. 
G1 is the number of genes, J is the number of samples.
}
\usage{data(matrixY1)}
\format{
  The format is:
 num [1:30, 1:15] 0.828 4.897 -0.609 2.856 0.497 ...
}
\examples{
data(matrixY1)
}
\keyword{datasets}
