\name{matrixY2}
\alias{matrixY2}
\docType{data}
\title{The drug sensitivity matrix}
\description{The drug sensitivity matrix, dim(Y2)=G2*J. 
G2 is the number of drugs, J is the number of samples.}
\usage{data(matrixY2)}
\format{
  The format is:
 num [1:30, 1:15] -0.157 -1.994 -0.339 -1.224 0.622 ...
}
\examples{
data(matrixY2)
}
\keyword{datasets}
