\name{matrixZ2}
\alias{matrixZ2}
\docType{data}
\title{Binary indictor matrix for matrixW2}
\description{The binary matrix indicating whether each entry of
    loading matrixW2 is non-zero. For example, matrixZ2[g,k]=1 indicates
    that matrixW2[g,k] is non-zero, and vice versa. Dim(matrixZ2)=G2*K.}
\usage{data(matrixZ2)}
\format{
  The format is:
 num [1:30, 1:10] 0 0 0 1 1 0 0 0 0 0 ...
}
\examples{
data(matrixZ2)
}
\keyword{datasets}
