\name{sigma1}
\alias{sigma1}
\docType{data}
\title{Covariance matrix of the noise term for the genes}
\description{The positive-definite symmetric matrix specifying 
   the covariance matrix of the noise term associated with each gene.
   Dim(sigma1)=G1*G1}
\usage{data(sigma1)}
\format{
  The format is:
 num [1:30, 1:30] 0.108 0 0 0 0 ...
}
\examples{
data(sigma1)
}
\keyword{datasets}
