\name{sigma2}
\alias{sigma2}
\docType{data}
\title{Covariance matrix of the noise term for the drugs}
\description{The positive-definite symmetric matrix specifying 
   the covariance matrix of the noise term associated with each drug.
   Dim(sigma2)=G2*G2}
\usage{data(sigma2)}
\format{
  The format is:
 num [1:30, 1:30] 0.00881 0 0 0 0 ...
}
\examples{
data(sigma2)
}
\keyword{datasets}
