\name{Y1_mean}
\alias{Y1_mean}
\docType{data}
\title{The mean value used for the simulation of matrixY1}
\description{The matrix of the mean value of matrixY1 before adding the
    noise term.Calculated by the multiplication of matrixW and matrixX.
    Dim(Y1_mean)=G1*J}
\usage{data(Y1_mean)}
\format{
  The format is:
 num [1:30, 1:15] 0.814 4.768 -0.475 2.84 0.579 ...
}
\examples{
data(Y1_mean)
}
\keyword{datasets}
