\name{Y2_mean}
\alias{Y2_mean}
\docType{data}
\title{The mean value used for the simulation of matrixY2}
\description{The matrix of the mean value of matrixY2 before adding the
    noise term.Calculated by the multiplication of matrixW and matrixX.
    Dim(Y2_mean)=G2*J}
\usage{data(Y2_mean)}
\format{
  The format is:
 num [1:30, 1:15] -0.138 -2.06 -0.242 -1.313 0.61 ...
}
\examples{
data(Y2_mean)
}
\keyword{datasets}
